+++ src/bn254/mod.rs
@@ -15,3 +15,4 @@ pub mod msm;
 pub mod utils;
 pub mod chunk_primitves;
 pub mod chunk_layout;
+pub mod chunk_taps;
